server/pyproject.toml
[project]
name = "cyberbuddha-x402"
version = "1.0.0"
description = "Cyber Buddha - 链上上香许愿 (x402)"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "x402>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.26.0",
]
workers = [
    "workers-py",
]
